requirements.txt
speechrecognition==3.10.0
pyttsx3==2.90
pyautogui==0.9.54
pygetwindow==0.0.9
google-generativeai>=0.3.0

requests>=2.31.0